clean up circular dependency between StructureComponent, StructureManager, GetComponent.
